Trendr/requirements.txt
scrapydo
tweepy
boto3
pytrends
requests
newsapi-python
vaderSentiment
scrapy